requirements.txt
appdirs==1.4.4
asgiref==3.2.10
boto==2.49.0
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-herokuapp==0.9.20
django-require==1.0.11
django-require-s3==1.0.0
django-storages==1.9.1
filelock==3.0.12
idna==2.10
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
python-dateutil==1.5
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
sh==1.13.1
waitress==1.4.4
virtualenv==20.0.30
whitenoise==5.2.0